Changed MessageContentActivity to not use fragment, disabled a preference
