+++ app/src/main/res/layout/content_message_content.xml
@@ -1,18 +1,21 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.kylemsguy.tcasmobile.MessageContentFragment">
+    app:layout_behavior="@string/appbar_scrolling_view_behavior"
+    tools:context="com.kylemsguy.tcasmobile.MessageContentActivity"
+    tools:showIn="@layout/activity_message_content">
 
     <TextView
         android:id="@+id/debug_msg_id"
-        android:text="@string/hello_world"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
+        android:layout_height="wrap_content"
+        android:text="@string/hello_world" />
 
     <android.support.v4.widget.SwipeRefreshLayout
         android:id="@+id/swipe_messages_refresh"
