+++ app/src/main/res/xml/pref_data_sync.xml
@@ -10,7 +10,8 @@
         android:entryValues="@array/pref_sync_frequency_values"
         android:defaultValue="180"
         android:negativeButtonText="@null"
-        android:positiveButtonText="@null" />
+        android:positiveButtonText="@null"
+        android:enabled="false"/>
 
     <!-- This preference simply launches an intent when selected. Use this UI sparingly, per
          design guidelines. -->
